Ignore any outbound requests for .corp
Microsoft organisations typically have a non-standard `.corp` at
the end of all hosts, which don't exist in the public internet.
Therefore there is no point in sending requests which end in `.corp`
to public internet servers.

If this is being used in a Microsoft environment, removing this file
may make sense.