Upstream DNS servers from Google and CloudFlare
When resolving names, a `server` will be used to forward request on,
either with IPv4 or IPv6 hosts. These files show how you could use
public DNS infrastructure provided by Google or CloudFlare to do the
upstream DNS provider.

It is expected that only one of these upstream servers sets will be
used, which can be achieved by deleting the other. However, it is
possible to use both at the same time, and requests will be
circulated around them.

Not shown is the use of `all-servers` which will broadcast a request
to all of them, and return the first one that responds. Having a
mixed strategy of a pair of Google and CloudFlare would thus return
the fastest result.