Test Docker image to use to validate concerns
When the docker image is built, it will create a copy of the dnsmasq
files and run a dnsmasq server based off of ubuntu:20.04. The build
process runs a `dnsmasq --test` phase, which will result in any
syntax errors being displayed and the Docker image not being built.

The docker image can be run to test out the setup, and by connecting
to it with a `docker exec -it dnsmasq bash` can perform tests to
verify the correctness of the setup before putting it live.

If the docker image is run with a volume mount, then interactive
tests can be performed simply by restarting the docker container
instead of requiring a rebuild.

    $ docker build . -t dnsmasq
    $ docker run --rm -it --name dnsmasq -v $(PWD)/dnsmasq.d:/etc/dnsmasq.d:ro dnsmasq
    $ docker exec -it dnsmasq dig @localhost example.com