Fix workaround to not return NXDOMAIN for .corp
The .corp address range is used by Microsoft products but shouldn't
ideally be forwarded upstream. However, returning NXDOMAIN caches
that value which then prevents Microsoft latops from being usable
once back on a VPN.

Fix this by returning a dummy NS (which will fail) so that the DNS
requests are not forwarded upstream but also which don't return a
value to the client which will be cached.