Add Cloudflare for Families DNS resolution
Cloudflare have release "Cloudflare for Families" which automatically
blocks DNS entries from hosts that have been classified as malware,
as porn hosting, or both.

One of the dns- entries should be kept, and the others ignored, in
order to take advantage of this.

See https://blog.cloudflare.com/introducing-1-1-1-1-for-families/
for more information.